refactor: remove StaticImage & change back to html img
With Gatsby, images are usually stored either in local storage (for e.g., the static folder) or on a hosted service/CMS. This requires images to be either 'hard coded' using StaticImage or for images with 'dynamic' sources, to be rendered using GatsbyImage. This requires the image url to be known at build time.

At this stage, it is unclear exactly what workaround would be needed for images that are added/created at run time in the browser. Thus, it would be easier to just switch back to using standard HTML img