refactor: image & label processing
The logic that took in the imageSrc, created a temporary Image object and used that object to determine the height and width of the uploaded image was moved into the PhotoAnalysis page. This was done to fix a build error stating that Image was not defined. The imageSrc data is not stored in a ImageData state